+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- N/A
+### Added
+
+- [Prisma ORM](https://www.prisma.io/) export format.
 
 ## [v8.6.1] - 2025-05-06